move fd_pack from ballet to disco
