+++ src/app/fdctl/run/tiles/fd_pack.c
@@ -5,8 +5,8 @@
 #include "../../../../disco/topo/fd_pod_format.h"
 #include "../../../../disco/shred/fd_shredder.h"
 #include "../../../../disco/metrics/fd_metrics.h"
-#include "../../../../ballet/pack/fd_pack.h"
-#include "../../../../ballet/pack/fd_pack_pacing.h"
+#include "../../../../disco/pack/fd_pack.h"
+#include "../../../../disco/pack/fd_pack_pacing.h"
 
 #include <linux/unistd.h>
 
@@ -682,7 +682,7 @@ during_frag( fd_pack_ctx_t * ctx,
 
     fd_txn_m_t * txnm = (fd_txn_m_t *)dcache_entry;
     fd_txn_t * txn  = fd_txn_m_txn_t( txnm );
-    
+
     fd_memcpy( ctx->cur_spot->txnp->payload, fd_txn_m_payload( txnm), txnm->payload_sz              );
     fd_memcpy( TXN(ctx->cur_spot->txnp),     txn,                     txnm->txn_t_sz                );
     fd_memcpy( ctx->cur_spot->alt_accts,     fd_txn_m_alut( txnm ),   32UL*txn->addr_table_adtl_cnt );
