+++ src/disco/pack/sample_vote.bin
