+++ src/disco/pack/test_chkdup.c
